ai_engine/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
torch
peft
sentencepiece
spacy>=3.7.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
textstat>=0.7.4
textblob>=0.17.0
nltk>=3.9
pydantic>=2.7.0
numpy>=1.26.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
python-dotenv>=1.0.1
